requirements.txt
streamlit  
pandas  
matplotlib  
statsmodels  
scikit-learn  
numpy  
seaborn  
plotly  
plotly.express  
plotly.graph_objects  
sklearn.linear_model  
sklearn.ensemble  
sklearn.tree  
sklearn.metrics  
sklearn.preprocessing  
sklearn.model_selection  
sklearn.decomposition  
warnings  
